db: make the postgres-only build (HL_ENABLE_SQLITE=0) link-clean
A SQLite-off, Postgres-on binary (HL_ENABLE_SQLITE=0 HL_ENABLE_POSTGRES=1)
failed to link: SQLite-specific code compiled under the HL_ENABLE_DB umbrella
referenced sqlite3_* / hl_db_sqlite_* unconditionally. Gate all of it on
HL_ENABLE_SQLITE so the 'any / neither backend' flag matrix is real.

- cap/db.c (the SQLite stmt-cache + query engine) is now filtered when
  HL_ENABLE_SQLITE=0. Its one backend-agnostic function,
  hl_cap_db_check_namespace (a pure _hull_* string check used by worker_db /
  mod_db), moved to cap/db_select.c, which is always compiled.
- db_backend.h: hl_db_backend_sqlite + the hl_db_sqlite_raw / _cache / _wrap /
  _unwrap accessors are gated on HL_ENABLE_SQLITE, with inline no-op fallbacks
  (return NULL / -1) so callers that legitimately degrade for a non-SQLite
  handle (app_context stmt cache, agent raw paths) stay compilable.
- mod_db.c (Lua + JS): the db.udf block (sqlite3_value / context marshalling,
  sqlite3_create_function) and the udf sub-object in push_conn_object are
  #ifdef HL_ENABLE_SQLITE. udf is inherently SQLite-only.
- worker_db.c: the 'realpath only when the backend is SQLite' check on
  &hl_db_backend_sqlite is gated.
- agent DB introspection (agent/db.c, sql.c, schema_diff.c, and
  hl_agent_open_app_db in helpers.c) opens the app's SQLite file directly, so
  it and its call sites (cmd_agent, mcp, agent_api) move from HL_ENABLE_DB to
  HL_ENABLE_SQLITE. The 'hull agent db|migrate|schema-diff|sql' subcommands
  are thus SQLite-only.

CI: the flavors matrix gains a 'postgres-only (no sqlite)' link check.
docs: CLAUDE.md + postgres_backend_design.md note the postgres-only build and
mark the follow-up done.

Validated: all four flavors build (default, sqlite-only, postgres-only,
sqlite+postgres) + 'hull version' runs; unit 54/54; db.udf + hull agent db
schema still work on SQLite; e2e_postgres passes.
