+++ src/hull/commands/agent.c
@@ -59,7 +59,7 @@ static int cmd_routes(int argc, char **argv)
     return output_result(&out, rc);
 }
 
-#ifdef HL_ENABLE_DB
+#ifdef HL_ENABLE_SQLITE
 static int cmd_db_schema(int argc, char **argv)
 {
     const char *app_dir = ".";
@@ -119,7 +119,7 @@ static int cmd_db(int argc, char **argv)
     fprintf(stderr, "hull agent db: unknown subcommand '%s'\n", argv[0]);
     return 1;
 }
-#endif /* HL_ENABLE_DB */
+#endif /* HL_ENABLE_SQLITE */
 
 static int cmd_request(int argc, char **argv)
 {
@@ -357,7 +357,7 @@ static int cmd_deploy(int argc, char **argv)
     return output_result(&out, rc);
 }
 
-#ifdef HL_ENABLE_DB
+#ifdef HL_ENABLE_SQLITE
 static int cmd_migrate(int argc, char **argv)
 {
     const char *app_dir = ".";
@@ -650,7 +650,7 @@ static int cmd_compute_call(int argc, char **argv)
     return output_result(&out, rc);
 }
 
-#ifdef HL_ENABLE_DB
+#ifdef HL_ENABLE_SQLITE
 static int cmd_schema_diff(int argc, char **argv)
 {
     const char *app_dir = ".";
@@ -748,7 +748,7 @@ int hl_cmd_agent(int argc, char **argv, const HlCommandEnv *env)
 
     if (strcmp(sub, "routes") == 0)
         return cmd_routes(sub_argc, sub_argv);
-#ifdef HL_ENABLE_DB
+#ifdef HL_ENABLE_SQLITE
     if (strcmp(sub, "db") == 0)
         return cmd_db(sub_argc, sub_argv);
 #endif
@@ -764,7 +764,7 @@ int hl_cmd_agent(int argc, char **argv, const HlCommandEnv *env)
         return cmd_test(sub_argc, sub_argv, env);
     if (strcmp(sub, "context") == 0)
         return cmd_context(sub_argc, sub_argv, env);
-#ifdef HL_ENABLE_DB
+#ifdef HL_ENABLE_SQLITE
     if (strcmp(sub, "migrate") == 0)
         return cmd_migrate(sub_argc, sub_argv);
 #endif
@@ -798,7 +798,7 @@ int hl_cmd_agent(int argc, char **argv, const HlCommandEnv *env)
     }
     /* Composite project summary — one call to orient an agent. */
     if (strcmp(sub, "overview") == 0)     return cmd_overview(sub_argc, sub_argv);
-#ifdef HL_ENABLE_DB
+#ifdef HL_ENABLE_SQLITE
     if (strcmp(sub, "schema-diff") == 0)  return cmd_schema_diff(sub_argc, sub_argv);
     if (strcmp(sub, "sql") == 0)          return cmd_sql(sub_argc, sub_argv);
 #endif
